Update about_page for modern ruby/rails
This gem was wildly out of date and needed updates in order to get the
test suite running and all tests passing. The biggest changes are using
modern sprockets conventions in the dummy app and updating how the
health check handles error messages as there was a large overhaul of how
ActiveModel.errors is populated around Rails 6.1.

In line with the changes to ActiveModel errors, the solr and fedora
configurations have been updated to add their errors in the modern way.

This commit also includes a number of linting fixes.